blog-microservice/requirements.txt
asgiref==3.7.2
boto3==1.34.59
botocore==1.34.59
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.4.4
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.2.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
